# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Book-A-Meal is an app designed to let you order meals from your favourite resturant at your convience. To checkout the app, visit [https://rovilay.github.io/Book-A-Meal/UI](https://rovilay.github.io/Book-A-Meal/UI) .",
    "version" : "1.0.0",
    "title" : "Book-A-Meal",
    "contact" : {
      "email" : "ogooluwaakinola@gmail.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    }
  },
  "host" : "book-me-a-meal.herokuapp.com",
  "basePath" : "/api/v1",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/api/v1/orders/<orderId>/meals" : {
      "get" : {
        "tags" : [ "Orders" ],
        "summary" : "Gets meals of an order",
        "description" : "This operation can be performed by admin(caterers) or the customer who placed the order",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "x-example" : "hg8f3183d-4bdf-4189-ac9e-af02a6d216c1"
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "x-example" : "0"
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "x-example" : "2"
        } ],
        "responses" : {
          "200" : {
            "description" : "'Order retrieved successfully!",
            "schema" : {
              "$ref" : "#/definitions/Order_Meals_Res"
            }
          },
          "400" : {
            "description" : "Error occurred while getting orders!",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "User not allowed!",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" :  "Order not found! OR Meal not found!",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        },
        "security" : [ {
          "jwtHeader" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "jwtHeader" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "Pagination" : {
      "properties" : {
        "limit" : {
          "type" : "integer",
          "format" : "int32"
        },
        "offset" : {
          "type" : "integer",
          "format" : "int32"
        },
        "count" : {
          "type" : "integer",
          "format" : "int32"
        },
        "numOfPages" : {
          "type" : "integer",
          "format" : "int32"
        },
        "curPage" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nextOffset" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "example" : {
        "numOfPages" : 5,
        "curPage" : 5,
        "offset" : 6,
        "limit" : 0,
        "count" : 1,
        "nextOffset" : 2
      }
    },
...
    "Order_User" : {
      "properties" : {
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        }
      },
      "description" : "User who places the order",
      "example" : {
        "firstName" : "firstName",
        "lastName" : "lastName"
      }
    },
...
    "OrderMeal" : {
      "properties" : {
        "portion" : {
          "type" : "integer",
          "format" : "int32"
        },
        "cost" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "example" : {
        "cost" : 5,
        "portion" : 1
      }
    },
    "order_Meals" : {
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "OrderMeal" : {
          "$ref" : "#/definitions/OrderMeal"
        }
      },
      "example" : {
        "OrderMeal" : {
          "cost" : 5,
          "portion" : 1
        },
        "id" : "id",
        "title" : "title"
      }
    },
    "order" : {
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "UserId" : {
          "type" : "string"
        },
        "deliveryAddress" : {
          "type" : "string"
        },
        "totalPrice" : {
          "type" : "integer",
          "format" : "int32"
        },
        "createdAt" : {
          "type" : "string"
        },
        "updatedAt" : {
          "type" : "string"
        },
        "User" : {
          "$ref" : "#/definitions/Order_User"
        },
        "Meals" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/order_Meals"
          }
        }
      },
      "example" : {
        "createdAt" : "createdAt",
        "User" : {
          "firstName" : "firstName",
          "lastName" : "lastName"
        },
        "deliveryAddress" : "deliveryAddress",
        "totalPrice" : 6,
        "UserId" : "UserId",
        "id" : "id",
        "Meals" : [ {
          "OrderMeal" : {
            "cost" : 5,
            "portion" : 1
          },
          "id" : "id",
          "title" : "title"
        }, {
          "OrderMeal" : {
            "cost" : 5,
            "portion" : 1
          },
          "id" : "id",
          "title" : "title"
        } ],
        "updatedAt" : "updatedAt"
      }
    },
    "Order_Meals_Res" : {
      "properties" : {
        "success" : {
          "type" : "boolean"
        },
        "message" : {
          "type" : "string"
        },
        "grandTotalPrice" : {
          "type" : "integer",
          "format" : "int32"
        },
        "pagination" : {
          "$ref" : "#/definitions/Pagination"
        },
        "order" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/order"
          }
        }
      },
      "example" : {
        "pagination" : {
          "numOfPages" : 5,
          "curPage" : 2,
          "offset" : 1,
          "limit" : 6,
          "count" : 5,
          "nextOffset" : 7
        },
        "success" : true,
        "message" : "message",
        "grandTotalPrice" : 0,
        "order" : [ {
          "createdAt" : "createdAt",
          "User" : {
            "firstName" : "firstName",
            "lastName" : "lastName"
          },
          "deliveryAddress" : "deliveryAddress",
          "totalPrice" : 6,
          "UserId" : "UserId",
          "id" : "id",
          "Meals" : [ {
            "OrderMeal" : {
              "cost" : 5,
              "portion" : 1
            },
            "id" : "id",
            "title" : "title"
          }, {
            "OrderMeal" : {
              "cost" : 5,
              "portion" : 1
            },
            "id" : "id",
            "title" : "title"
          } ],
          "updatedAt" : "updatedAt"
        }, {
          "createdAt" : "createdAt",
          "User" : {
            "firstName" : "firstName",
            "lastName" : "lastName"
          },
          "deliveryAddress" : "deliveryAddress",
          "totalPrice" : 6,
          "UserId" : "UserId",
          "id" : "id",
          "Meals" : [ {
            "OrderMeal" : {
              "cost" : 5,
              "portion" : 1
            },
            "id" : "id",
            "title" : "title"
          }, {
            "OrderMeal" : {
              "cost" : 5,
              "portion" : 1
            },
            "id" : "id",
            "title" : "title"
          } ],
          "updatedAt" : "updatedAt"
        } ]
      }
    },
...
    "Error" : {
      "properties" : {
        "success" : {
          "type" : "boolean",
          "default" : false
        },
        "message" : {
          "type" : "string"
        }
      },
      "description" : "Response recieved when operation is unsuccessful"
    }
  },
...
}